svg-uri removed from Collapsible/2, styles moved to constants files, ChevronLeftBlack created
