+++ components/Animals/AnimalCard.js
@@ -3,7 +3,7 @@ import { View, Text, TouchableOpacity, Image } from 'react-native';
 import * as WebBrowser from 'expo-web-browser';
 import { AmpEvent } from '../../components/withAmplitude';
 
-import ChevronLeft from '../../assets/jsicons/miscIcons/ChevronLeftSolid';
+import ChevronLeft from '../../assets/jsicons/miscIcons/ChevronLeftWhite';
 
 import styles from '../../constants/Animals/AnimalCard';
